fix null data from backend causing runtime error
